+++ jax/lax/lax.py
@@ -24,7 +24,6 @@ from typing import Any
 import warnings
 
 import numpy as onp
-from scipy.special import factorial as fact  # TODO scipy dep?
 
 from ..util import partial, prod
 
@@ -44,7 +43,6 @@ from ..interpreters import xla
 from ..interpreters import pxla
 from ..interpreters import ad
 from ..interpreters import batching
-from ..interpreters import taylor
 from ..interpreters import masking
 from ..util import curry, cache, safe_zip, unzip2, prod
 from ..tree_util import build_tree, tree_unflatten, tree_map
@@ -1640,7 +1638,6 @@ _bool_or_int = _int | _bool
 
 neg_p = standard_unop(_num, 'neg')
 ad.deflinear(neg_p, lambda t: [neg(t)])
-taylor.deflinear(neg_p)
 
 def _sign_translation_rule(c, x):
   shape = c.GetShape(x)
@@ -1674,33 +1671,8 @@ ad.defjvp_zero(is_finite_p)
 exp_p = standard_unop(_float | _complex, 'exp')
 ad.defjvp2(exp_p, lambda g, ans, x: _safe_mul(g, ans))
 
-def _exp_taylor(primals_in, series_in):
-  x, = primals_in
-  series, = series_in
-  u = [x] + series
-  v = [exp(x)] + [None] * len(series)
-  def scale(k, j): return 1. / (fact(k-j) * fact(j-1))
-  for k in range(1,len(v)):
-    v[k] = fact(k-1) * sum([scale(k, j)* v[k-j] * u[j] for j in range(1, k+1)])
-  primal_out, *series_out = v
-  return primal_out, series_out
-taylor.jet_rules[exp_p] = _exp_taylor
-
 log_p = standard_unop(_float | _complex, 'log')
 ad.defjvp(log_p, lambda g, x: div(g, x))
-
-def _log_taylor(primals_in, series_in):
-  x, = primals_in
-  series, = series_in
-  u = [x] + series
-  v = [log(x)] + [None] * len(series)
-  def scale(k, j): return 1. / (fact(k-j) * fact(j-1))
-  for k in range(1, len(v)):
-    conv = sum([scale(k, j) * v[j] * u[k-j] for j in range(1, k)])
-    v[k] = (u[k] - fact(k - 1) * conv) / u[0]
-  primal_out, *series_out = v
-  return primal_out, series_out
-taylor.jet_rules[log_p] = _log_taylor
 
 expm1_p = standard_unop(_float | _complex, 'expm1')
 ad.defjvp2(expm1_p, lambda g, ans, x: mul(g, add(ans, _one(ans))))
@@ -1792,7 +1764,6 @@ ad.defjvp2(erf_inv_p, lambda g, ans, x: mul(_const(x, onp.sqrt(onp.pi) / 2.),
 
 real_p = unop(_complex_basetype, _complex, 'real')
 ad.deflinear(real_p, lambda t: [complex(t, onp.zeros((), _dtype(t)))])
-taylor.deflinear(real_p)
 
 imag_p = unop(_complex_basetype, _complex, 'imag')
 ad.defjvp(imag_p, lambda g, _: real(mul(_const(g, -1j), g)))
@@ -1801,7 +1772,6 @@ _complex_dtype = lambda dtype, *args: (onp.zeros((), dtype) + onp.zeros((), onp.
 complex_p = naryop(_complex_dtype, [_complex_elem_types, _complex_elem_types],
                   'complex')
 ad.deflinear(complex_p, lambda t: [real(t), imag(neg(t))])
-taylor.deflinear(complex_p)
 
 conj_p = unop(_complex_dtype, _complex_elem_types | _complex, 'conj')
 
@@ -1869,7 +1839,6 @@ def _add_transpose(t, x, y):
 add_p = standard_naryop([_num, _num], 'add')
 ad.defjvp(add_p, lambda g, x, y: _brcast(g, y), lambda g, x, y: _brcast(g, x))
 ad.primitive_transposes[add_p] = _add_transpose
-taylor.deflinear(add_p)
 
 
 def _sub_transpose(t, x, y):
@@ -1880,7 +1849,6 @@ sub_p = standard_naryop([_num, _num], 'sub')
 ad.defjvp(sub_p,
           lambda g, x, y: _brcast(g, y),
           lambda g, x, y: _brcast(neg(g), x))
-taylor.deflinear(sub_p)
 ad.primitive_transposes[sub_p] = _sub_transpose
 
 mul_p = standard_naryop([_num, _num], 'mul')
@@ -1908,21 +1876,7 @@ div_p = standard_naryop([_num, _num], 'div')
 ad.defjvp(div_p,
           lambda g, x, y: div(_brcast(g, y), y),
           lambda g, x, y: div(mul(neg(_brcast(g, x)), x), square(y)))
-
-def _div_taylor_rule(primals_in, series_in, **params):
-  x, y = primals_in
-  x_terms, y_terms = series_in
-  u = [x] + x_terms
-  w = [y] + y_terms
-  v = [None] * len(u)
-  def scale(k, j): return 1. / (fact(k - j) * fact(j))
-  for k in range(0, len(v)):
-    conv = sum([scale(k, j) * v[j] * w[k-j] for j in range(0, k)])
-    v[k] = (u[k] - fact(k) * conv) / w[0]
-  primal_out, *series_out = v
-  return primal_out, series_out
 ad.primitive_transposes[div_p] = _div_transpose_rule
-taylor.jet_rules[div_p] = _div_taylor_rule
 
 rem_p = standard_naryop([_num, _num], 'rem')
 ad.defjvp(rem_p,
@@ -2012,7 +1966,6 @@ convert_element_type_p = standard_primitive(
 ad.deflinear(
     convert_element_type_p,
     lambda t, new_dtype, old_dtype: [convert_element_type(t, old_dtype)])
-taylor.deflinear(convert_element_type_p)
 batching.defvectorized(convert_element_type_p)
 masking.defvectorized(convert_element_type_p)
 
@@ -2372,22 +2325,6 @@ ad.defbilinear(dot_general_p,
 batching.primitive_batchers[dot_general_p] = _dot_general_batch_rule
 masking.masking_rules[dot_general_p] = _dot_general_masking_rule
 
-def _bilinear_taylor_rule(prim, primals_in, series_in, **params):
-  x, y = primals_in
-  x_terms, y_terms = series_in
-  u = [x] + x_terms
-  w = [y] + y_terms
-  v = [None] * len(u)
-  op = partial(prim.bind, **params)
-  def scale(k, j): return 1. / (fact(k - j) * fact(j))
-  for k in range(0, len(v)):
-    v[k] = fact(k) * sum([scale(k, j) * op(u[j], w[k-j]) for j in range(0, k+1)])
-  primal_out, *series_out = v
-  return primal_out, series_out
-taylor.jet_rules[dot_general_p] = partial(_bilinear_taylor_rule, dot_general_p)
-taylor.jet_rules[mul_p] = partial(_bilinear_taylor_rule, mul_p)
-taylor.jet_rules[conv_general_dilated_p] = partial(_bilinear_taylor_rule, conv_general_dilated_p)
-
 
 def _broadcast_shape_rule(operand, sizes):
   _check_shapelike('broadcast', 'sizes', sizes)
@@ -2402,7 +2339,6 @@ def _broadcast_batch_rule(batched_args, batch_dims, sizes):
 broadcast_p = standard_primitive(
     _broadcast_shape_rule, _input_dtype, 'broadcast')
 ad.deflinear(broadcast_p, lambda t, sizes: [_reduce_sum(t, range(len(sizes)))])
-taylor.deflinear(broadcast_p)
 batching.primitive_batchers[broadcast_p] = _broadcast_batch_rule
 
 def _broadcast_in_dim_impl(operand, shape, broadcast_dimensions):
@@ -2450,7 +2386,6 @@ broadcast_in_dim_p = standard_primitive(
     _broadcast_in_dim_shape_rule, _input_dtype, 'broadcast_in_dim')
 broadcast_in_dim_p.def_impl(_broadcast_in_dim_impl)
 ad.deflinear(broadcast_in_dim_p, _broadcast_in_dim_transpose_rule)
-taylor.deflinear(broadcast_in_dim_p)
 batching.primitive_batchers[broadcast_in_dim_p] = _broadcast_in_dim_batch_rule
 
 
@@ -2541,7 +2476,6 @@ concatenate_p = standard_primitive(
     _concatenate_shape_rule, _concatenate_dtype_rule, 'concatenate',
     _concatenate_translation_rule)
 ad.deflinear(concatenate_p, _concatenate_transpose_rule)
-taylor.deflinear(concatenate_p)
 ad.primitive_transposes[concatenate_p] = _concatenate_transpose_rule
 batching.primitive_batchers[concatenate_p] = _concatenate_batch_rule
 
@@ -2590,7 +2524,6 @@ def _pad_batch_rule(batched_args, batch_dims, padding_config):
 
 pad_p = standard_primitive(_pad_shape_rule, _pad_dtype_rule, 'pad')
 ad.deflinear(pad_p, _pad_transpose)
-taylor.deflinear(pad_p)
 ad.primitive_transposes[pad_p] = _pad_transpose
 batching.primitive_batchers[pad_p] = _pad_batch_rule
 
@@ -2683,9 +2616,7 @@ reshape_p = standard_primitive(_reshape_shape_rule, _reshape_dtype_rule,
                                'reshape', _reshape_translation_rule)
 reshape_p.def_impl(_reshape_impl)
 ad.deflinear2(reshape_p, _reshape_transpose_rule)
-taylor.deflinear(reshape_p)
 batching.primitive_batchers[reshape_p] = _reshape_batch_rule
-taylor.deflinear(reshape_p)
 
 
 def _rev_shape_rule(operand, dimensions):
@@ -2707,7 +2638,6 @@ def _rev_batch_rule(batched_args, batch_dims, dimensions):
 
 rev_p = standard_primitive(_rev_shape_rule, _input_dtype, 'rev')
 ad.deflinear(rev_p, lambda t, dimensions: [rev(t, dimensions)])
-taylor.deflinear(rev_p)
 batching.primitive_batchers[rev_p] = _rev_batch_rule
 
 
@@ -2740,7 +2670,6 @@ transpose_p = standard_primitive(_transpose_shape_rule, _input_dtype,
 transpose_p.def_impl(_transpose_impl)
 ad.deflinear(transpose_p,
              lambda t, permutation: [transpose(t, onp.argsort(permutation))])
-taylor.deflinear(transpose_p)
 batching.primitive_batchers[transpose_p] = _transpose_batch_rule
 
 
@@ -2899,7 +2828,6 @@ def _slice_batching_rule(batched_args, batch_dims, start_indices, limit_indices,
 slice_p = standard_primitive(_slice_shape_rule, _input_dtype, 'slice',
                              _slice_translation_rule)
 ad.deflinear2(slice_p, _slice_transpose_rule)
-taylor.deflinear(slice_p)
 batching.primitive_batchers[slice_p] = _slice_batching_rule
 
 
@@ -3198,14 +3126,6 @@ gather_p = standard_primitive(
     _gather_shape_rule, _gather_dtype_rule, 'gather',
     _gather_translation_rule)
 ad.defjvp(gather_p, _gather_jvp_rule, None)
-
-def _gather_taylor_rule(primals_in, series_in, **params):
-  operand, start_indices = primals_in
-  gs, _ = series_in
-  primal_out = gather_p.bind(operand, start_indices, **params)
-  series_out = [gather_p.bind(g, start_indices, **params) for g in gs]
-  return primal_out, series_out
-taylor.jet_rules[gather_p] = _gather_taylor_rule
 
 ad.primitive_transposes[gather_p] = _gather_transpose_rule
 batching.primitive_batchers[gather_p] = _gather_batching_rule
@@ -3557,7 +3477,6 @@ reduce_sum_p = standard_primitive(
   _reduce_sum_shape_rule, partial(_reduce_number_dtype_rule, 'reduce_sum'),
   'reduce_sum', _reduce_sum_translation_rule)
 ad.deflinear2(reduce_sum_p, _reduce_sum_transpose_rule)
-taylor.deflinear(reduce_sum_p)
 batching.defreducer(reduce_sum_p)
 _masking_defreducer(reduce_sum_p,
                     lambda shape, dtype: onp.broadcast_to(onp.array(0, dtype), shape))
@@ -3633,23 +3552,6 @@ _reduce_max_translation_rule = partial(_reduce_chooser_translation_rule, max_p,
 reduce_max_p = standard_primitive(_reduce_op_shape_rule, _input_dtype,
                                   'reduce_max', _reduce_max_translation_rule)
 ad.defjvp2(reduce_max_p, _reduce_chooser_jvp_rule)
-
-def _reduce_max_taylor_rule(primals_in, series_in, **params):
-  operand, = primals_in
-  gs, = series_in
-  primal_out = reduce_max_p.bind(operand, **params)
-  axes = params.pop("axes", None)
-  primal_dtype = gs[0].dtype
-  shape = [1 if i in axes else d for i, d in enumerate(operand.shape)]
-  location_indicators = convert_element_type(
-        _eq_meet(operand, reshape(primal_out, shape)), primal_dtype)
-  counts = _reduce_sum(location_indicators, axes)
-  def _reduce_chooser_taylor_rule(g):
-    return div(_reduce_sum(mul(g, location_indicators), axes), counts)
-  series_out = [_reduce_chooser_taylor_rule(g) for g in gs]
-  return primal_out, series_out
-taylor.jet_rules[reduce_max_p] = _reduce_max_taylor_rule
-
 batching.defreducer(reduce_max_p)
 
 
@@ -3776,7 +3678,6 @@ reduce_window_sum_p = standard_primitive(
     _reduce_window_sum_shape_rule, _input_dtype, 'reduce_window_sum',
     _reduce_window_sum_translation_rule)
 ad.deflinear2(reduce_window_sum_p, _reduce_window_sum_transpose_rule)
-taylor.deflinear(reduce_window_sum_p)
 batching.primitive_batchers[reduce_window_sum_p] = partial(
   _reduce_window_batch_rule, _reduce_window_sum)
 
@@ -4234,7 +4135,6 @@ tie_in_p.def_impl(lambda x, y: y)
 tie_in_p.def_abstract_eval(lambda x, y: raise_to_shaped(y))
 xla.translations[tie_in_p] = lambda c, x, y: y
 ad.deflinear(tie_in_p, _tie_in_transpose_rule)
-taylor.deflinear(tie_in_p)
 batching.primitive_batchers[tie_in_p] = _tie_in_batch_rule
 masking.shape_rules[tie_in_p] = lambda x, y: y.shape
 masking.masking_rules[tie_in_p] = lambda vals, logical_shapes: vals[1]
